+++ services/web/src/API/Dashboard/MissingEpisodes/Get/Route.hs
@@ -1,0 +1,20 @@
+module API.Dashboard.MissingEpisodes.Get.Route where
+
+--------------------------------------------------------------------------------
+
+import Domain.Types.Cookie (Cookie)
+import Domain.Types.HxRequest (HxRequest)
+import Lucid qualified
+import Servant ((:>))
+import Servant qualified
+import Text.HTML (HTML)
+
+--------------------------------------------------------------------------------
+
+-- | "GET /dashboard/missing-episodes"
+type Route =
+  "dashboard"
+    :> "missing-episodes"
+    :> Servant.Header "Cookie" Cookie
+    :> Servant.Header "HX-Request" HxRequest
+    :> Servant.Get '[HTML] (Lucid.Html ())